ACC set fuel

+++ src/PitCrew.Interface/Pipeline/IPitStrategyRequest.cs
@@ -2,5 +2,5 @@
 
 public interface IPitStrategyRequest
 {
-    int FuelToAdd { get; }
+    double FuelToAdd { get; }
 }
